Add files via upload
This document captures Mike’s 1st bootstrap learning of Pipedream, an integration platform that I am going to use to route data from the Helium Console to Google Sheets.
This document goes through a step-by-step analysis and copy of a 4 minute Pipedream YouTube video demonstrating how to Analyze Twitter Sentiment in Real-Time and to save the results to a Google Sheet. If you follow the steps in this document, you should have your own working solution for Twitter-to-Google Sheets.